+++ SQLiteBrowser/TableDetailPage.cs
@@ -61,8 +61,16 @@ public partial class TableDetailPage : ContentPage
                     TextColor = Color.FromRgb(0, 0, 0),
                     IsVisible = HasRows,
                     Margin = new Thickness(0, 10, 0, 0)
+                },
+                new Button
+                {
+                    Text = "Delete table (DROP)",
+                    Command = new Command(OnDropTableBtnClicked),
+                    BackgroundColor = Color.FromRgb(255, 0, 0),
+                    IsVisible = HasRows,
+                    Margin = new Thickness(0, 10, 0, 0)
                 }
-                
+
             }
         };
     }
@@ -154,6 +162,21 @@ public partial class TableDetailPage : ContentPage
         try
         {
             SQLiteWrapper.Truncate(this.dbPath, this.tableName);
+
+            await Navigation.PopAsync();
+        }
+        catch (Exception ex)
+        {
+            await DisplayAlert("Error", ex.Message, "OK");
+        }
+    }
+
+    private async void OnDropTableBtnClicked(object obj)
+    {
+        if (!await DisplayAlert("Confirm", "Are you sure you want to DROP?", "Yes", "No")) return;
+        try
+        {
+            SQLiteWrapper.Drop(this.dbPath, this.tableName);
 
             await Navigation.PopAsync();
         }